requirements.txt
aiofile==3.7.4
anyio==3.4.0
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
caio==0.9.3
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysqlclient==2.1.0
oauthlib==3.1.1
orjson==3.6.5
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.0
requests-oauthlib==1.3.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
SQLAlchemy-ImageAttach==1.1.0
starlette==0.16.0
toml==0.10.2
tweepy==4.4.0
typing-extensions==4.0.1
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
Wand==0.6.7
watchgod==0.7
websockets==10.1
